image loading complete

+++ berlin/preprocess/audio.py
@@ -67,7 +67,7 @@ def pool_to_features(pool, config):
         array = essentia.array(values)
         array = array[config.frames_start_index:config.frames_end_index]
         array = (array - np.mean(array)) / np.std(array)
-        result[key] = array
+        result[key] = array.astype('float16')
 
     return Features(**result)
